tehtava 4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -65,15 +65,41 @@ blogsRouter.post('/', async (request, response, next) => {
 // REMOVE ONE
 blogsRouter.delete('/:id', async (request, response, next) => {
   logger.info("delete request")
+
+  const token = request.token
+
+  let decodedToken
   try {
-    const deleted = await Blog.findByIdAndDelete(request.params.id)
-    if(deleted){
-      response.status(204).end()
-    } else {
-      logger.info("id not found - returning 400")
-      response.status(400).end()
-    }
+    decodedToken = jwt.verify(token, process.env.SECRET)
   } catch(e) {
+    next(e)
+  }
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  // Haetaan lähettänyt käyttäjä
+  const user = await User.findById(decodedToken.id)
+
+  try {
+    // Haetaan poistettava blogi id-polkuparametrin avulla
+    const blogToBeDeleted = await Blog.findById(request.params.id)
+    if(blogToBeDeleted){
+      // Tarkastetaan vastaako poistettavan blogin lähettäjä poistoehdotuksen lähettäjää
+      if(blogToBeDeleted.user.toString() != user._id.toString()){
+        logger.info("Tried to delete someone elses blog")
+        return response.status(401).json({ error: 'token not corresponding the original poster' })
+      } else {
+        const deleted = await Blog.findByIdAndDelete(request.params.id)   
+        logger.info("succesfully deleted")
+        response.status(204).end()
+      }
+  } else {
+    logger.info("id not found - returning 400")
+    response.status(400).end()
+  }
+}
+  catch(e) {
     next(e)
   }
 })
